+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Not yet released
+* Allow building with `template-haskell-2.15` by manually implementing
+  `Lift` for `Bytes`. See [#27]
+
+[#27]: https://github.com/mgsloan/th-orphans/issues/27
+
 ### 0.13.6 [2018.07.01]
 * Allow building with `template-haskell-2.14`.
 * Implement `qAddForeignFilePath` and `qAddTempFile` for `Quasi` instances